Humayon

+++ src/app/modules/blog/blog.service.ts
@@ -4,22 +4,85 @@ import QueryBuilder from '../../builder/QueryBuilder';
 import { TPackage } from '../package/package.interface';
 import { access, unlink } from 'fs/promises';
 import Blog from './blog.model';
+import { deleteFromS3, deleteManyFromS3, uploadManyToS3, uploadToS3 } from '../../utils/s3';
+import { TBlog } from './blog.interface';
 
-const createBlog = async (payload: TPackage) => {
+const createBlog = async (files: any, payload: TBlog) => {
   try {
+
+
+    if (files.image && files.image.length > 0) {
+      const image: any = await uploadToS3({
+        file: files.image[0],
+        fileName: files.image[0].originalname,
+        folder: 'blogs/',
+      });
+      payload.image = image;
+    }
+    // if (files.bodyImage && files.bodyImage.length > 0) {
+    //   const image: any = await uploadToS3({
+    //     file: files.bodyImage[0],
+    //     fileName: files.bodyImage[0].originalname,
+    //     folder: 'blogs/',
+    //   });
+    //   payload.bodyImage = image;
+    // }
+    // if (files.upload3Photos && files.upload3Photos.length > 0) {
+    //   const upload3Photos = await uploadManyToS3(files.upload3Photos, 'blogs/');
+    //   payload.upload3Photos = upload3Photos;
+    // }
+    // if (files.ugcImage && files.ugcImage.length > 0) {
+    //   const image: any = await uploadToS3({
+    //     file: files.ugcImage[0],
+    //     fileName: files.ugcImage[0].originalname,
+    //     folder: 'blogs/',
+    //   });
+    //   payload.ugcImage = image;
+    // }
+
+
     const result = await Blog.create(payload);
+    if (result) {
+      const fileDeletePathimage = `${files.image[0].path}`;
+      await unlink(fileDeletePathimage);
+
+      // const fileDeletePathbodyImage = `${files.bodyImage[0].path}`;
+      // await unlink(fileDeletePathbodyImage);
+
+      // const fileDeletePathupload3Photos =files.upload3Photos.map((file: any) => `${file.path}`);
+      // for (const fileDeletePath of fileDeletePathupload3Photos) {
+      //   await unlink(fileDeletePath);
+      // }
+
+      // const fileDeletePathugcImage = `${files.ugcImage[0].path}`;
+      // await unlink(fileDeletePathugcImage);
+      
+    }
     return result;
   } catch (error) {
     try {
-      const imagePath = `public/${payload.image}`;
-      await access(imagePath);
-      await unlink(imagePath);
+      const fileDeletePathimage = `${files.image[0].path}`;
+      await unlink(fileDeletePathimage);
+
+      // const fileDeletePathbodyImage = `${files.bodyImage[0].path}`;
+      // await unlink(fileDeletePathbodyImage);
+
+      // const fileDeletePathupload3Photos = files.upload3Photos.map(
+      //   (file: any) => `${file.path}`,
+      // );
+      // for (const fileDeletePath of fileDeletePathupload3Photos) {
+      //   await unlink(fileDeletePath);
+      // }
+
+      // const fileDeletePathugcImage = `${files.ugcImage[0].path}`;
+      // await unlink(fileDeletePathugcImage);
     } catch (fsError) {
       console.error('Error accessing or deleting the image file:', fsError);
     }
     throw error;
   }
 };
+
 
 const getAllBlogQuery = async (query: Record<string, unknown>) => {
   const blogQuery = new QueryBuilder(Blog.find(), query)
@@ -43,7 +106,7 @@ const getSingleBlogQuery = async (id: string) => {
   return blog;
 };
 
-const updateSingleBlogQuery = async (id: string, payload: any) => {
+const updateSingleBlogQuery = async (id: string, files: any, payload: any) => {
   try {
     console.log('id', id);
     console.log('updated payload', payload);
@@ -54,39 +117,48 @@ const updateSingleBlogQuery = async (id: string, payload: any) => {
       throw new AppError(404, 'Package not found!');
     }
 
-    // Update package
-    const updatedPackage = await Blog.findByIdAndUpdate(id, payload, {
-      new: true,
-    });
-    if (!updatedPackage) {
-      throw new AppError(403, 'Package update failed!');
-    }
+    if (files?.image && files?.image.length > 0) {
+      const image: any = await uploadToS3({
+        file: files.image[0],
+        fileName: files.image[0].originalname,
+        folder: 'blogs/',
+      });
+      payload.image = image;
 
-    console.log('result', updatedPackage);
-
-    // Only delete old image if a new image is provided in the payload and it's different
-    if (payload.image && payload.image !== existingPackage.image) {
-      const oldImagePath = `public/${existingPackage.image}`;
-      console.log('Deleting old image at:', oldImagePath);
-
-      try {
-        await access(oldImagePath);
-        await unlink(oldImagePath);
-      } catch (fsError) {
-        console.error('Error accessing or deleting old image file:', fsError);
+      const result = await Blog.findByIdAndUpdate(id, payload, {
+        new: true,
+      });
+      if (result) {
+        const fileDeletePath = `${files.image[0].path}`;
+        await unlink(fileDeletePath);
       }
+
+      const key = existingPackage.image.split('amazonaws.com/')[1];
+
+      const deleteImage: any = await deleteFromS3(key);
+      console.log('deleteImage', deleteImage);
+      if (!deleteImage) {
+        throw new AppError(404, 'Blog Image Deleted File !');
+      }
+
+      return result;
+
+
+    }else{
+
+      const result = await Blog.findByIdAndUpdate(id, payload, {
+        new: true,
+      });
+      return result;
+
     }
 
-    return updatedPackage;
   } catch (error) {
-    if (payload.image) {
-      const newImagePath = `public/${payload.image}`;
-      try {
-        await access(newImagePath);
-        await unlink(newImagePath);
-      } catch (fsError) {
-        console.error('Error accessing or deleting new image file:', fsError);
-      }
+    try {
+      const fileDeletePath = `${files.image[0].path}`;
+      await unlink(fileDeletePath);
+    } catch (fsError) {
+      console.error('Error accessing or deleting the image file:', fsError);
     }
     throw error;
   }
@@ -101,12 +173,32 @@ const deletedBlogQuery = async (id: string) => {
     throw new AppError(404, 'Blog Not Found!!');
   }
 
-  const result = await Blog.findByIdAndDelete(id);
-  if (!result) {
-    throw new AppError(404, 'Blog Result Not Found !');
+  console.log('blogs', blog)
+
+  const key1 = blog.image.split('amazonaws.com/')[1];
+  const deleteImage1:any = await deleteFromS3(key1);
+
+  // const key2 = blog.bodyImage.split('amazonaws.com/')[1];
+  // const deleteImage2:any = await deleteFromS3(key2);
+
+  // const key3 = blog.ugcImage.split('amazonaws.com/')[1];
+  // const deleteImage3:any = await deleteFromS3(key3);
+
+
+  // const keys = blog?.upload3Photos?.map(
+  //   (key: any) => key.url.split('amazonaws.com/')[1],
+  // );
+  // const deleteImages: any = await deleteManyFromS3(keys);
+  // console.log('deleteImage', deleteImages);
+
+  if (deleteImage1 ) {
+    const result = await Blog.findByIdAndDelete(id);
+    return result;
+  } else {
+    throw new AppError(500, 'Failed to delete one or more images from S3');
   }
 
-  return result;
+
 };
 
 export const blogService = {
